requirements.txt
requests
pyyaml
logger
pydantic